Moved payload for sending GlobalAnomalyStats from pserver->viz to class header
